refactor: update type -> instance_type, size is float type, datetime to iso8601

+++ src/spaceone/inventory/connector/aws_s3_connector/connector.py
@@ -45,7 +45,7 @@ class S3Connector(SchematicAWSConnector):
                         'name': data.name,
                         'data': data,
                         'account': self.account_id,
-                        'size': int(data.size),
+                        'size': float(data.size),
                         'reference': ReferenceModel(data.reference())
                     }
